Move quiz's package types to types.go
